Moved primitives into one file
