+++ src/index.ts
@@ -1,8 +1,1 @@
-export { blob } from '@/src/model/primitives/blob';
-export { boolean } from '@/src/model/primitives/boolean';
-export { date } from '@/src/model/primitives/date';
-export { json } from '@/src/model/primitives/json';
-export { number } from '@/src/model/primitives/number';
-export { link } from '@/src/model/primitives/link';
-export { string } from '@/src/model/primitives/string';
-export { model } from '@/src/model/model';
+export * from '@/src/model/primitives';
