--- src/model/primitives/blob.ts
@@ -1,19 +1,0 @@
-import type { SerializedField } from '@/src/model/model';
-
-/**
- * Creates a blob field definition returning an object that includes the field type
- * ("blob") and attributes.
- *
- * @param attributes - Optional field attributes.
- *
- * @returns A field of type "blob" with the specified or default attributes.
- */
-export const blob = (attributes: SerializedField<'blob'> = {}) => {
-  const { name, ...rest } = attributes;
-
-  return {
-    name,
-    type: 'blob' as const,
-    ...rest,
-  };
-};
